FEAT(If the F key is pressed the fullscreen mode is reversed), FEAT(resize window, reload map)

+++ include/game.h
@@ -25,7 +25,7 @@ typedef struct game {
     map_t map;
     //player_t player;
     object_t *duck;
-    double speedx;
+    float speedx;
     double posx;
 } game_t;
